menambahkan produk, serta fitur tambah, update dan delete

+++ adminpanel/kategori-detail.php
@@ -69,6 +69,18 @@
                 }
 
                 if(isset($_POST['deleteBtn'])){
+                    $queryCheck = mysqli_query($con, "SELECT * FROM produk WHERE kategori_id = $id");
+                    $dataCount = mysqli_num_rows($queryCheck);
+                    
+                    if($dataCount > 0){
+                        ?>
+                            <div class="alert alert-warning mt-3" role="alert">
+                                Kategori tidak bisa dihapus karena memiliki produk
+                            </div>
+                        <?php
+                        die();
+                    }
+
                     $queryDelete = mysqli_query($con, "DELETE FROM kategori WHERE id='$id'");
                     if($queryDelete){
                         ?>
